backend/requirements.txt
fastapi[standard]
langchain-community
langchainhub 
langgraph
langchain-openai